# scaffold/hr-service/openapi.yaml
openapi: 3.0.3
info:
  title: PMP Human Resources Service
  version: "1.0"
  description: >
    Implements FR-HR-01..07 (docs Section 4.6). FastAPI must generate a schema
    that matches this contract — treat mismatches as a bug, not a spec update,
    unless the SRS itself is being revised.

    All endpoints require a bearer access token (RS256 JWT) issued by
    iam-service; its signature is verified against iam-service's JWKS and the
    required `<resource>.<action>` permission must be present in the token's
    `permissions` claim (CLAUDE.md rule 4). Missing/invalid token -> 401;
    valid token without the permission -> 403.

    discipline_records are additionally restricted: only hr.discipline.read /
    hr.discipline.write unlock those routes (docs Section 2.3 — HR/command
    only). FR-HR-08 (roster/headcount reports) is not implemented — it reads
    to derive reports rather than defining new domain state, and is deferred.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /officers/{officer_id}/assignments:
    post:
      summary: Record a unit assignment (FR-HR-02)
      description: >
        Requires `hr.assignment.write`. Closes the officer's currently open
        assignment (if any) and updates their current unit. Publishes an
        `AssignmentRecorded` domain event.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: officer_id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssignmentCreate" }
      responses:
        "201": { description: Recorded, content: { application/json: { schema: { $ref: "#/components/schemas/Assignment" } } } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.assignment.write }
        "404": { description: No officer with that id, or unit_id does not exist }
    get:
      summary: Assignment history for one officer (FR-HR-02)
      description: Requires `hr.assignment.read`.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: officer_id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Assignments, newest first
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Assignment" } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.assignment.read }
        "404": { description: No officer with that id }
...
  /officers/{officer_id}/discipline-records:
    post:
      summary: Record a discipline case (FR-HR-06)
      description: >
        Requires `hr.discipline.write`. Publishes a `DisciplineRecordCreated`
        domain event.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: officer_id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DisciplineRecordCreate" }
      responses:
        "201": { description: Recorded, content: { application/json: { schema: { $ref: "#/components/schemas/DisciplineRecord" } } } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.discipline.write }
        "404": { description: No officer with that id }
    get:
      summary: Discipline-record history for one officer (FR-HR-06)
      description: >
        Requires `hr.discipline.read` — HR/command only (docs Section 2.3).
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: officer_id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Discipline records, newest first
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/DisciplineRecord" } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.discipline.read }
        "404": { description: No officer with that id }
  /discipline-records/{discipline_record_id}:
    get:
      summary: Retrieve one discipline record (FR-HR-06)
      description: >
        Requires `hr.discipline.read` — HR/command only (docs Section 2.3).
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: discipline_record_id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/DisciplineRecord" } } } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.discipline.read }
        "404": { description: No discipline record with that id }
    patch:
      summary: Update a discipline record (FR-HR-06)
      description: >
        Requires `hr.discipline.write`. Publishes a `DisciplineRecordUpdated`
        domain event.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: discipline_record_id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DisciplineRecordUpdate" }
      responses:
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/DisciplineRecord" } } } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.discipline.write }
        "404": { description: No discipline record with that id }
    delete:
      summary: Delete a discipline record (FR-HR-06)
      description: >
        Requires `hr.discipline.write`. Publishes a `DisciplineRecordDeleted`
        domain event.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: discipline_record_id, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Deleted }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.discipline.write }
        "404": { description: No discipline record with that id }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 access token issued by iam-service (/api/v1/auth/mfa/verify).
  schemas:
...
    AssignmentCreate:
      type: object
      required: [unit_id, start_date]
      properties:
        unit_id: { type: string, format: uuid }
        start_date: { type: string, format: date }
    Assignment:
      allOf:
        - $ref: "#/components/schemas/AssignmentCreate"
        - type: object
          properties:
            id: { type: string, format: uuid }
            officer_id: { type: string, format: uuid }
            end_date: { type: string, format: date, nullable: true }
...
    DisciplineRecordCreate:
      type: object
      properties:
        confidentiality_level: { type: string, default: restricted }
    DisciplineRecord:
      allOf:
        - $ref: "#/components/schemas/DisciplineRecordCreate"
        - type: object
          properties:
            id: { type: string, format: uuid }
            officer_id: { type: string, format: uuid }
            created_at: { type: string, format: date-time }
    DisciplineRecordUpdate:
      type: object
      properties:
        confidentiality_level: { type: string, nullable: true }
...